Add prototype with working video.
This is very much a work in progress and is not yet functioning. Video is being read to the GUI
and image comparisons are working, but the program is not yet assembled in a way that makes it
usable. The biggest obstacle is making the backend and frontend work together.