Add GUI controller
GUI controller is still under construction, and the program is broken again while
I'm setting it up. There were too many signals flying back and forth between classes
that were supposed to add up to determining the GUI logic, and the code was becoming
too difficult to maintain. There's still a lot of work to do.