Fully implement GUI controller
The GUI controller works now. It is much easier to handle button and label settings this way.
I haven't tested everything (split flags, for one), but it seems like things are largely working
as expected. Next step is to make some minor GUI adjustments and fully implement the settings window
and hotkeys.