Update settings menu, add persistent storage
Finished the threading implementation -- hopefully there won't be any issues
with crashing. I found one bug and (to my best knowledge) quashed it -- apparently
calling QPixmap.isNull() is not thread-safe, at least not on my machine, which is
weird. Still lots to do, but the settings menu is 90% implemented and working great.
This program should be more or less usable in its current state.