Roll split_image, split_directory into splitter
Still reworking the structure of the project. Next step is to figure out the flow.
The idea right now is to have a timer call capture.get_frame every 1/FPS seconds,
within splitter I think, and then I guess I will have splitter process the output?
I also want to have the capture send frames directly to the UI. Not sure how to handle
this cleanly yet, I need to sleep on it.

I am also continuing to work on shifting large portions of main_window and settings_window
to the controller to make things easier to maintain. This is a work in progress, and this branch's
code is completely broken.