Configure build settings for MacOS
Added build.sh, which compiles a MacOS executible using Pyinstaller
and injects it into a Platypus application, along with a bootstrap
script that starts it. Apparently it's infamously difficult to get a
GUI app in Python to run on MacOS, so this is as good as it's going
to get, I think. It's compiling fine right now (after lots of annoyances).