Make settings.py check for and return specific types
Apparently, depending on the backend QSettings may only return strings.
I've reworked settings.py to include various getter methods, which return
str, int, float, bool, and QKeySequence values; and a single setter method,
which saves the value as a string. This should be compatible with any backend.

Also fixed a bug that was preventing match_percent from updating correctly on
the UI after switching to a new image folder (again!).