Fix compare thread moving ahead 2 images
This is the last time I fix this bug, I think. The issue is that when
pynput makes a key press, there is a slight delay (sometimes as much as ~.2
sec but usually around .01 sec) before PyQt recognizes the key press.
This made simply setting a flag to prevent the key press from splitting
difficult. I've added a thread, so now when the splitter presses the split
key, it waits 1 second in a separate thread before unsetting the flag,
which should always be enough time to prevent a double split.

Also moved hotkey back to its own module.