Add flag to Poller for reset split labels only
Last commit, I handled this by directly setting the split image's pixmap,
but there is more than one situation that calls for resetting the split image
even when the splitter hasn't done anything unusual (e.g. pressing the reset
button). So I added a flag to Poller that is set when the aspect ratio changes
or when the reset button / hotkey is pressed.