Move set_layout, others to ui_controller
It was too complicated to figure out all the methods that were setting
flags to trigger methods in ui_controller. To simplify things, I moved
set_layout to ui_controller (as well as a couple other methods). Now,
for the most part, all information that is pulled from splitter to UI
is handled completely in ui_controller, which is exactly what I want.