Force compare thread to always use 320x240 images
Splitter._capture now stores two kinds of frames: (1) ui_frame,
which is downscaled to the size of the ui video feed using the
fastest, poorest-quality method available, and which is fed to the
ui; (2) comparison_frame, which is always downscaled to 320x240, uses
a slightly better downscaling algorithm, and is only used in Splitter._compare.

Split images are also stored in two sizes: the 320x240 size SplitImage.image,
which is fed to Splitter._compare and uses the worst-quality downscaling
algorithm, and the ui size SplitImage.pixmap, which uses the same downscaling
algorithm as comparison_frame.

This has the effect of bringing the CPU load on all ui sizes down to the level
of the smallest UI size, which is fantastic.

I also made a minor adjustment to the way ui.controller's request_previous_split
and request_next_split work. Instead of restarting Splitter._compare_thread, which
is slow, they spawn a new thread that watches for a change in the current split index
or loop, then sets current_match_percent and highest_match_percent to 0. This is much
faster and allows for "fast scrolling" (e.g. holding down the next split key until
you get to the end of the split list). The thread times out after 1 second, so it
won't hang around if something goes wrong.