Implement global hotkeys with pynput.keyboard
splitter.py:

Get rid of _split_thread, because with global hotkeys it is no
longer needed. It was needed since there was always a chance that
the split hotkey would exist, be activated, but not be heard by the
program (e.g. if the program was not in focus). That is not a risk
anymore. Also add a change_splits block to _compare (see below).

ui_controller.py:

Get rid of match_percent_helper_thread, because it was causing problems and
was buggy, and replace it with a simpler system -- the "change_splits"
block in splitter._compare. Now in request_next_split and request_previous_split,
the controller waits for _compare to send back a "waiting" signal, changes the
split image, then restarts _compare without killing the thread. This keeps
those two methods fast but allows all flags to reset (including below_flag
and went_above_threshold_flag, which were giving me trouble).

Implement hotkey listener. When the settings window is opened, it
changes the setting of the hotkey line edit that is in focus, if there
is one. When it's closed, it listens for hotkey presses and sets flags,
which are picked up and acted on by Poller. I'm not sure if I love that
Poller is playing this roll, since I feel like Poller's purpose should
be to gather information from Splitter, not from user input, but it's
good enough for now.

Always restart compare thread when reset button / shortcut is pressed, so
that the splitter's suspend / delay status is cleared.

Name screenshots starting from 000, not 001.

ui_settings_window.py:

Open this window with open(), not exec(). This allows me to get rid of
a bunch of redundant keyboard shortcuts I implemented just to make this
window easier to close.

Set is_opened flag when opening and closing the window, so that hotkeys
are not sent to the splitter when typing in new settings, etc.

ui_main_window.py:

Update help link