Allow disabling global hotkeys
The UI controller checks if global hotkeys are enabled, or if one of the program windows is in focus, before executing a hotkey action after a hotkey is pressed. Hotkey presses (e.g. when a split happens) are disabled if global hotkeys are disabled.

Other changes:

Transferred control of split execution (pressing the actual hotkey) from splitter.py to ui_controller.py. ui_controller already handles everything else when it comes to hotkeys, so it made sense to just move everything there. Now, splitter.py sets a flag when a split occurs, which ui_controller reads and acts on.

Made settings_window open with exec() again instead of open(). exec() blocks the main window and I want that behavior while the settings window is open. Relatedly, I overrode exec() and got rid of closeEvent() and open() in settings_window, so everything is handled through exec(), which is cleaner.

Rewrote ui_controller.save_settings for readability and maintainability