Handle grayscale (1 channel) images in split_dir
Add _is_single_channel, which is called when making images and pixmaps.
If an image is grayscale, it's converted to a BGR image. This should almost
never happen, but it's good to handle it.

Also removed some extraneous code from splitter and split_dir that isn't needed
now that comparison images are always the same size, regardless of aspect ratio.