Fix icon not showing sometimes, Python error msg
Fixed icon not showing in taskbar when app is run from anywhere
besides the app's root directory

Fixed (mostly) the "Python quit unexpectedly." error that pops up
frequently when quitting the app. It still happens once in a while,
but not as often (I think). There might be more I can do to prevent
it from popping up if it is still annoying after a while.