Fix splitter thread startup issue
I had changed ui_controller._request_reset_splits so that it checked
if the splitter was active in order to decide whether to start it again
after resetting, but I think it's more intuitive to start the splitter
if splits and video are both loaded. If the user doesn't want that, they
can press "Pause comparison" after pressing Reset splits.