Enable global hotkeys on startup (sometimes)
If the user has a hotkey bound for toggling global hotkeys, I think it's
safe to assume that they use global hotkeys and toggle regularly. If
that's the case, it makes sense to set global hotkeys enabled on startup
if they're not enabled already (lest there be confusion when the user
opens the app, expecting global hotkeys to work, only to discover they're
off because they quit the app after disabling them last time). On the
other hand, if the user has global hotkeys disabled and no key bound to
enable them, they're probably just not interested, so we can leave them
off.