Allow binding same hotkey to multiple actions
This will likely rarely be useful, but if, for example, someone
wanted to take a screenshot on every split, you could bind the
same key to both actions.

Also added a timeout to the popup messages raised by _take_screenshot,
and switched from using exec() to raise them to using show(), since
show() is non-blocking. Now if someone accidentally or intentionally
takes a screenshot during a run, the program will continue functioning
as it should.