Always respond to "toggle globals" hotkey
Previously I had this only work (if global hotkeys were off) when the
app was in focus, but sometimes it can be cumbersome to click back into
the app just to reenable global hotkeys (especially when you're using
something like LiveSplit One, where you have to make sure it's in focus
because it doesn't have global hotkeys). Choosing an out-of-the-way key
should make this easy to work around.