Improve performance of SplitImage.get_split_images
Since get_split_images is called every time the reset button / reset
hotkey is pressed, it needs to be fast. Now instead of making each split
image one at a time, a threadpool is dispatched to make them in parallel.
In addition, instead of naively making a split image object for each image,
we check if an object in the same index, with the same path, and with the
same last modified time exists, and if so, we use that object instead.
These two changes together reduce the reset time by about 100x on my
machine when there are about 100 split images. The gains for smaller lists
are also very significant.