Transer update checker to main_window
Also removed the setting I added earlier that turns global hotkeys
back on on launch if a hotkey is bound (see commit #858b984). I
feel like it's too hands-on.

When the "Update Available" message box is shown directly from the
main window, it behaves very differently than it was behaving before.
I'm not sure why, but that's why there are so many changes to main_window.