Shorten varnames, save CPU in ui_controller
Shortened variable names throughout the program, especially ones that
are frequently used. Also, I removed many of the if/ then constructs
from the code in ui_controller that update the UI. To my surprise,
this dramatically reduced CPU usage (and as an added bonus, the code
is much more readable). I wanted to simplify that part of ui_controller
because its unnecessarily complex structure makes it very difficult
to maintain, and I want to make some adjustments.