Lower FPS min to 10, set 20Hz min for poller
I find that an FPS setting of half your game's FPS is sufficient for
accurate splitting. Since there are 20 FPS games (like the N64 Zelda
games), it makes sense to allow the user to set FPS to 10. In light
of this, I also introduced a 20Hz floor for ui_controller._poller,
because when the application is only refreshing at 10Hz it looks very
choppy.