Remove delay/pause overlay from min view
I feel like the purpose of minimal view is to keep distractions to
a minimum while providing necesssary information. My purpose in
showing delay/pause in min view was to help troubleshoot if you're
trying to figure out if pauses are too long / short, but that kind
of troubleshooting can be done outside of min view. I think things
are good enough as they are, and I find the extra overlay to be
distracting.