Fix occasional crash when trying to bind hotkeys
Previously, if the user was typing very quickly with a hotkey box in
focus in the settings menu, a segmentation fault could rarely occur.
This has been fixed: _handle_hotkey_press now sets flags instead of
modifying the GUI directly (something it has done for a long time to
press hotkeys, but didn't seem necessary for simply updating hotkey
values). A lock is used to avoid racing when updating the box.